pull mcap stream matching out of grid previews
